+++ apps/api/src/billing/services/provider-cleanup/provider-cleanup.service.ts
@@ -4,16 +4,11 @@ import { singleton } from "tsyringe";
 import { BillingConfig, InjectBillingConfig } from "@src/billing/providers";
 import { UserWalletOutput, UserWalletRepository } from "@src/billing/repositories";
 import { ManagedUserWalletService, RpcMessageService } from "@src/billing/services";
+import { ProviderCleanupParams } from "@src/billing/types/provider-cleanup";
 import { ErrorService } from "@src/core/services/error/error.service";
 import { ProviderCleanupSummarizer } from "@src/deployment/lib/provider-cleanup-summarizer/provider-cleanup-summarizer";
 import { DeploymentRepository } from "@src/deployment/repositories/deployment/deployment.repository";
 import { TxSignerService } from "../tx-signer/tx-signer.service";
-
-export interface ProviderCleanupParams {
-  concurrency: number;
-  provider: string;
-  dryRun: boolean;
-}
 
 @singleton()
 export class ProviderCleanupService {