+++ apps/api/src/deployment/services/top-up-custodial-deployments/top-up-custodial-deployments.service.spec.ts
@@ -1,8 +1,8 @@
 import "@test/mocks/logger-service.mock";
 
 import { AllowanceHttpService, BalanceHttpService, Denom } from "@akashnetwork/http-sdk";
-import { faker } from "@faker-js/faker";
 import { MsgExec } from "cosmjs-types/cosmos/authz/v1beta1/tx";
+import { secondsInWeek } from "date-fns/constants";
 import { describe } from "node:test";
 import { container } from "tsyringe";
 
@@ -24,6 +24,8 @@ import { DeploymentGrantSeeder } from "@test/seeders/deployment-grant.seeder";
 import { DrainingDeploymentSeeder } from "@test/seeders/draining-deployment.seeder";
 import { FeesAuthorizationSeeder } from "@test/seeders/fees-authorization.seeder";
 import { stub } from "@test/services/stub";
+
+const USDC_IBC_DENOM = "ibc/170C677610AC31DF0904FFE09CD3B5C657492170E7E52372E48756B71E56F2F1";
 
 describe(TopUpCustodialDeploymentsService.name, () => {
   const CURRENT_BLOCK_HEIGHT = 7481457;
@@ -71,13 +73,14 @@ describe(TopUpCustodialDeploymentsService.name, () => {
 
   type SeedParams = {
     denom: Denom;
-    balance?: string;
+    balance?: number;
+    feesBalance?: number;
     grantee: string;
     expectedDeploymentsTopUpCount?: 0 | 1 | 2;
     hasDeployments?: boolean;
   };
 
-  const seedFor = ({ denom, balance = "100000000", grantee, expectedDeploymentsTopUpCount = 2, hasDeployments = true }: SeedParams) => {
+  const seedFor = ({ denom, balance = 100000000, feesBalance = 1000000, grantee, expectedDeploymentsTopUpCount = 2, hasDeployments = true }: SeedParams) => {
     const owner = AkashAddressSeeder.create();
 
     return {
@@ -90,8 +93,9 @@ describe(TopUpCustodialDeploymentsService.name, () => {
       feeAllowance: FeesAuthorizationSeeder.create({
         granter: owner,
         grantee: grantee,
-        allowance: { spend_limit: { denom } }
+        allowance: { spend_limit: { denom: "uakt" } }
       }),
+      feesBalance: denom === "uakt" ? undefined : BalanceSeeder.create({ denom: "uakt", amount: feesBalance }),
       drainingDeployments: hasDeployments
         ? [
             {
@@ -113,24 +117,66 @@ describe(TopUpCustodialDeploymentsService.name, () => {
       grantee: UAKT_TOP_UP_MASTER_WALLET_ADDRESS
     }),
     seedFor({
-      denom: "ibc/170C677610AC31DF0904FFE09CD3B5C657492170E7E52372E48756B71E56F2F1",
+      denom: USDC_IBC_DENOM,
       grantee: USDT_TOP_UP_MASTER_WALLET_ADDRESS
     }),
     seedFor({
+      denom: USDC_IBC_DENOM,
+      grantee: USDT_TOP_UP_MASTER_WALLET_ADDRESS,
+      balance: 5500000,
+      expectedDeploymentsTopUpCount: 2
+    }),
+    seedFor({
+      denom: USDC_IBC_DENOM,
+      grantee: USDT_TOP_UP_MASTER_WALLET_ADDRESS,
+      balance: 5040000,
+      expectedDeploymentsTopUpCount: 1
+    }),
+    seedFor({
+      denom: USDC_IBC_DENOM,
+      grantee: USDT_TOP_UP_MASTER_WALLET_ADDRESS,
+      balance: 5500000,
+      expectedDeploymentsTopUpCount: 2
+    }),
+    seedFor({
+      denom: USDC_IBC_DENOM,
+      grantee: USDT_TOP_UP_MASTER_WALLET_ADDRESS,
+      feesBalance: 0,
+      expectedDeploymentsTopUpCount: 0
+    }),
+    seedFor({
+      denom: USDC_IBC_DENOM,
+      grantee: USDT_TOP_UP_MASTER_WALLET_ADDRESS,
+      feesBalance: 5000,
+      expectedDeploymentsTopUpCount: 1
+    }),
+    seedFor({
       denom: "uakt",
-      balance: "5500000",
+      balance: 5045000,
       grantee: UAKT_TOP_UP_MASTER_WALLET_ADDRESS,
       expectedDeploymentsTopUpCount: 1
     }),
     seedFor({
       denom: "uakt",
-      balance: "5500000",
+      balance: 5000,
+      grantee: UAKT_TOP_UP_MASTER_WALLET_ADDRESS,
+      expectedDeploymentsTopUpCount: 0
+    }),
+    seedFor({
+      denom: "uakt",
+      balance: 10000,
+      grantee: UAKT_TOP_UP_MASTER_WALLET_ADDRESS,
+      expectedDeploymentsTopUpCount: 1
+    }),
+    seedFor({
+      denom: "uakt",
+      balance: 5500000,
       grantee: UAKT_TOP_UP_MASTER_WALLET_ADDRESS,
       hasDeployments: false
     }),
     seedFor({
       denom: "uakt",
-      balance: "0",
+      balance: 0,
       grantee: UAKT_TOP_UP_MASTER_WALLET_ADDRESS,
       expectedDeploymentsTopUpCount: 0
     })
@@ -143,12 +189,20 @@ describe(TopUpCustodialDeploymentsService.name, () => {
     return data.find(({ grant }) => grant.granter === granter && grant.grantee === grantee)?.feeAllowance;
   });
   jest.spyOn(balanceHttpService, "getBalance").mockImplementation(async (address: string, denom: Denom) => {
-    return (
-      data.find(({ grant }) => grant.granter === address)?.balance || {
-        amount: "0",
-        denom
-      }
-    );
+    const record = data.find(({ grant }) => grant.granter === address);
+
+    if (record?.balance.denom === denom) {
+      return record.balance;
+    }
+
+    if (record?.feesBalance.denom === denom) {
+      return record.feesBalance;
+    }
+
+    return {
+      amount: 0,
+      denom
+    };
   });
   jest.spyOn(drainingDeploymentService, "findDeployments").mockImplementation(async (owner, denom) => {
     return (
@@ -157,18 +211,22 @@ describe(TopUpCustodialDeploymentsService.name, () => {
         ?.drainingDeployments?.map(({ deployment }) => deployment) || []
     );
   });
-  jest.spyOn(drainingDeploymentService, "calculateTopUpAmount").mockImplementation(async () => faker.number.int({ min: 3500000, max: 4000000 }));
+  jest.spyOn(drainingDeploymentService, "calculateTopUpAmount").mockImplementation(async ({ blockRate }) => (blockRate * secondsInWeek) / 6);
 
   it("should top up draining deployment given owners have sufficient grants and balances", async () => {
     await topUpDeploymentsService.topUpDeployments({ dryRun: false });
 
-    expect(uaktMasterSigningClientService.executeTx).toHaveBeenCalledTimes(3);
-    expect(usdtMasterSigningClientService.executeTx).toHaveBeenCalledTimes(2);
+    let uaktCount = 0;
+    let usdtCount = 0;
 
     data.forEach(({ drainingDeployments, grant }) => {
       drainingDeployments.forEach(({ isExpectedToTopUp, deployment }) => {
         if (isExpectedToTopUp) {
-          const client = deployment.denom === "uakt" ? uaktMasterSigningClientService : usdtMasterSigningClientService;
+          const isAkt = deployment.denom === "uakt";
+          const client = isAkt ? uaktMasterSigningClientService : usdtMasterSigningClientService;
+          uaktCount += isAkt ? 1 : 0;
+          usdtCount += isAkt ? 0 : 1;
+
           expect(client.executeTx).toHaveBeenCalledWith(
             [
               {
@@ -189,6 +247,9 @@ describe(TopUpCustodialDeploymentsService.name, () => {
         }
       });
     });
+
+    expect(uaktMasterSigningClientService.executeTx).toHaveBeenCalledTimes(uaktCount);
+    expect(usdtMasterSigningClientService.executeTx).toHaveBeenCalledTimes(usdtCount);
   });
 
   xdescribe("actual top up deployment tx on demand", () => {