+++ apps/api/src/deployment/types/state-deployments.ts
@@ -1,0 +1,3 @@
+import { ConcurrencyOptions } from "@src/core/types/console";
+
+export interface CleanUpStaleDeploymentsParams extends ConcurrencyOptions {}