+++ apps/deploy-web/src/components/wallet/ConnectManagedWalletButton.tsx
@@ -15,8 +15,14 @@ export const ConnectManagedWalletButton: React.FunctionComponent<Props> = ({ cla
   const { connectManagedWallet, hasManagedWallet, isWalletLoading } = useWallet();
 
   return (
-    <Button variant="outline" onClick={connectManagedWallet} className={cn("border-primary bg-primary/10 dark:bg-primary", className)} {...rest} disabled={isWalletLoading}>
-      {isWalletLoading ? <Spinner size="small" className="mr-2" /> : <Rocket className="text-xs" />}
+    <Button
+      variant="outline"
+      onClick={connectManagedWallet}
+      className={cn("border-primary bg-primary/10 dark:bg-primary", className)}
+      {...rest}
+      disabled={isWalletLoading}
+    >
+      {isWalletLoading ? <Spinner size="small" className="mr-2" variant="dark" /> : <Rocket className="rotate-45 text-xs" />}
       <span className="m-2 whitespace-nowrap">{hasManagedWallet ? "Switch to USD Payments" : "Start Trial"}</span>
     </Button>
   );